src/AIAgent/requirements.txt
# =============================================================================
# AI Agent — Python Dependencies
# Azure AI Foundry SDK + Evaluation + OpenTelemetry
# =============================================================================
azure-ai-projects>=2.1.0
azure-identity>=1.19.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.27.0
openai>=1.58.0
python-dotenv>=1.0.1
